precision issue confirmed
notice that when binaries are not exactly 0 or 1 then conditions are
relaxed quite a bit because of large M values. Decreasing the tolerance
causes the little relaxation which means valid solutions are found. M
values should be kept less than tolerance, ex. M=1e7 and binary tol=1e-7
works, but M=1e9 and binary tol = 1e-5 does not work.